Extract CryptoResolver
